requirements.txt
asgiref==3.10.0
attrs==25.4.0
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
djangorestframework==3.16.1
fastf1==3.6.1
fonttools==4.60.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
requests-cache==1.2.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
timple==0.1.8
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
websockets==13.1
python-decouple==3.8
dj-database-url==3.0.1
django-cors-headers==4.3.1
